+++ csocket.h
@@ -49,7 +49,7 @@ class CSocket {
 
     // network measurements/estimates for this connection.
     u_long bandwidth();
-    u_long RTT();
+    double RTT();
     ssize_t trickle_chunksize(struct timeval time_since_last_fg,
                               struct timeval bg_wait_time);
   private:
